+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change venue when match warmup starts
 - New endpoint `/tournament/<id>/matches/result` returning results for all matches in tournament
 - Added more general statistics for each tournament to `/tournament/<id>/statistics`
+- Added visit statistics for 60+, 100+, 140+, and 180
 
 #### Changed
 - Switched from string to time for datetimes to correctly support timezones